move GsParameters interface
