+++ application/src/ext-test/java/org/opentripplanner/ext/datastore/gs/GsDataSourceRepositoryTest.java
@@ -6,6 +6,7 @@ import static org.junit.jupiter.api.Assertions.assertNull;
 import java.net.URI;
 import org.junit.jupiter.api.Test;
 import org.opentripplanner.datastore.api.FileType;
+import org.opentripplanner.datastore.api.GsParameters;
 
 public class GsDataSourceRepositoryTest {
