+++ application/src/ext/java/org/opentripplanner/ext/datastore/gs/GsDataSourceRepository.java
@@ -12,6 +12,7 @@ import java.util.Collections;
 import org.opentripplanner.datastore.api.CompositeDataSource;
 import org.opentripplanner.datastore.api.DataSource;
 import org.opentripplanner.datastore.api.FileType;
+import org.opentripplanner.datastore.api.GsParameters;
 import org.opentripplanner.datastore.base.DataSourceRepository;
 import org.opentripplanner.datastore.file.ZipStreamDataSourceDecorator;
