+++ application/src/main/java/org/opentripplanner/datastore/api/OtpDataStoreConfig.java
@@ -5,7 +5,6 @@ import java.util.List;
 import java.util.regex.Pattern;
 import javax.annotation.Nullable;
 import org.opentripplanner.datastore.OtpDataStore;
-import org.opentripplanner.ext.datastore.gs.GsParameters;
 
 /**
  * The {@link OtpDataStore} config, needed to create a store. This should be provided by the OTP
@@ -19,13 +18,7 @@ public interface OtpDataStoreConfig {
   URI reportDirectory();
 
   /**
-   * Local file system path to Google Cloud Platform service accounts credentials file. The
-   * credentials is used to access GCS urls. When using GCS from outside of the bucket cluster you
-   * need to provide a path the the service credentials.
-   * <p>
-   * This is a path to a file on the local file system, not an URI.
-   * <p>
-   * Optional. May return {@code null}.
+   * Parameters used to connect to Google Cloud Storage.
    */
   GsParameters gsParameters();
